requirements.txt
selenium==4.1.0
robotframework
robotframework-appiumlibrary
robotframework-seleniumlibrary